chore: clean up of utils
