--- src/interfaces/metaData.ts
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-export interface MetaData {
-  prcgTmDp: number;
-  prcgTmCRSP: number;
-  traceParent?: string | null;
-}
